+++ autoclass/__init__.py
@@ -1,8 +1,9 @@
 from autoclass.autoargs import *
+from autoclass.autoclass import *
 from autoclass.autoprops import *
 from autoclass.autodict import *
 from autoclass.var_checker import *
 from autoclass.validate import *
-from autoclass.booleans import *
+from autoclass.utils_typing import *
 
-__all__ = ['autoargs', 'autoprops', 'autodict', 'validate', 'var_checker', 'booleans']
+__all__ = ['autoclass', 'autoargs', 'autoprops', 'autodict', 'validate', 'var_checker', 'utils_typing']